src/functions/requirements.txt
functions-framework==3.*
firebase-admin==6.2.0
firebase-functions==0.5.0
google-cloud-firestore==2.13.0
flask==2.3.3
flask-cors==6.0.2
